rename MiniReviewable to UserMenuReviewable
